Add admin privileges to modifyPost, deletePost & deleteComment in controllers

+++ backend/controllers/comment.controller.js
@@ -9,7 +9,7 @@ async function createComment(req, res, next) {
     let infoComment = {
       user_id: req.user,
       post_id: req.params.post_id,
-      ...req.body
+      ...req.body,
     };
     console.log(infoComment);
     const newComment = await Comment.create(infoComment);
@@ -40,7 +40,8 @@ async function modifyComment(req, res, next) {
 async function deleteComment(req, res, next) {
   try {
     const commentObject = await Comment.findOne({
-      where: { comment_id: req.params.id }, include: [User]
+      where: { comment_id: req.params.id },
+      include: [User],
     });
 
     // if (!commentObject) {
@@ -51,10 +52,17 @@ async function deleteComment(req, res, next) {
     //   res.status(400).send({ message: "Unauthorized request !" });
     // }
 
-    await Comment.destroy({
-      where: { comment_id: req.params.id },
-    });
-    res.status(200).send({ message: "Comment deleted !" });
+    if (commentObject.user_id == req.user || req.admin == true) {
+      await Comment.destroy({
+        where: { comment_id: req.params.id },
+      });
+      res.status(200).send({ message: "Comment deleted !" });
+    } else {
+      res.status(401).json({
+        message:
+          "You are not the post owner or you do not have root privileges !",
+      });
+    }
   } catch (error) {
     res.status(500).json({ message: error.message });
     next();
@@ -64,7 +72,8 @@ async function deleteComment(req, res, next) {
 async function getOneComment(req, res, next) {
   try {
     const oneComment = await Comment.findOne({
-      where: { post_id: req.params.post_id, comment_id: req.params.id }, include: [User, Post]
+      where: { post_id: req.params.post_id, comment_id: req.params.id },
+      include: [User, Post],
     });
     res.status(200).send(oneComment);
   } catch (error) {
@@ -74,7 +83,10 @@ async function getOneComment(req, res, next) {
 
 async function getAllComment(req, res, next) {
   try {
-    const allComment = await Comment.findAll({ where: { post_id: req.params.post_id }, include: [User, Post] });
+    const allComment = await Comment.findAll({
+      where: { post_id: req.params.post_id },
+      include: [User, Post],
+    });
     res.status(200).send(allComment);
   } catch (error) {
     res.status(500).json({ message: error.message });
